// packages/api/src/gen/sast/sastshopv2/catalog/v1/product_template_pb.ts
// @generated by protoc-gen-es v2.12.0 with parameter "target=ts"
// @generated from file sast/sastshopv2/catalog/v1/product_template.proto (package sast.sastshopv2.catalog.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Timestamp } from "@bufbuild/protobuf/wkt";
import { file_google_protobuf_timestamp } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file sast/sastshopv2/catalog/v1/product_template.proto.
 */
export const file_sast_sastshopv2_catalog_v1_product_template: GenFile = /*@__PURE__*/
  fileDesc("CjFzYXN0L3Nhc3RzaG9wdjIvY2F0YWxvZy92MS9wcm9kdWN0X3RlbXBsYXRlLnByb3RvEhpzYXN0LnNhc3RzaG9wdjIuY2F0YWxvZy52MSLBAQoPUHJvZHVjdFRlbXBsYXRlEgoKAmlkGAEgASgDEg0KBXRpdGxlGAIgASgJEhMKC2Rlc2NyaXB0aW9uGAMgASgJEhMKC3ByaWNlX2NlbnRzGAQgASgFEhAKCHN0b3JlX2lkGAUgASgDEhYKDm1haW5faW1hZ2VfdXJsGAYgASgJEg8KB2JhcmNvZGUYByABKAkSLgoKdXBkYXRlZF9hdBgIIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXBiBnByb3RvMw", [file_google_protobuf_timestamp]);

/**
 * @generated from message sast.sastshopv2.catalog.v1.ProductTemplate
 */
export type ProductTemplate = Message<"sast.sastshopv2.catalog.v1.ProductTemplate"> & {
  /**
   * @generated from field: int64 id = 1;
   */
  id: bigint;

  /**
   * @generated from field: string title = 2;
   */
  title: string;

  /**
   * @generated from field: string description = 3;
   */
  description: string;

  /**
   * @generated from field: int32 price_cents = 4;
   */
  priceCents: number;

  /**
   * @generated from field: int64 store_id = 5;
   */
  storeId: bigint;

  /**
   * @generated from field: string main_image_url = 6;
   */
  mainImageUrl: string;

  /**
   * @generated from field: string barcode = 7;
   */
  barcode: string;

  /**
   * @generated from field: google.protobuf.Timestamp updated_at = 8;
   */
  updatedAt?: Timestamp | undefined;
};

/**
 * Describes the message sast.sastshopv2.catalog.v1.ProductTemplate.
 * Use `create(ProductTemplateSchema)` to create a new message.
 */
export const ProductTemplateSchema: GenMessage<ProductTemplate> = /*@__PURE__*/
  messageDesc(file_sast_sastshopv2_catalog_v1_product_template, 0);